Insert/pyproject.toml
[project]
name = "insert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "httpx[socks]>=0.28.1",
    "lxml>=6.0.2",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
